refactor: clean up transfer function
